committing my progress for as a save 2026-04-29
