+++ quizzes/src/quiz08/b_urie/Cargo.toml
@@ -3,10 +3,6 @@ name = "b_urie"
 version = "0.1.0"
 edition = "2024"
 
-[[bin]]
-name = "panic"
-path = "src/main.rs"
-
 [dependencies]
 chrono = "0.4.38"
 tokio = { version = "1.52.0", features = ["macros", "rt-multi-thread"] }
